This is a temporary patch for SDL to allow it to run XCSoar on the KOBO Mini
It does the following:

1) Force framebuffer
2) Force double buffering
3) Dither as framebuffer is blitted to screen to be black and white.  This uses a
   randomly changing dither offset to help remove ghosts.  Black and white is required
   for the high speed updates on the eink screen
4) Use event0/event1 for keyboard and mouse

It doesn't have any support for setting up the screen or the ioctls needed to
refresh the e-ink display.  This is a HACK IN PROGRESS and really should be moved
into a seperate driver rather than fbconns.  I was more interested in flying than doing it properly