From 779022bfdc6c203e8a423ece480e8799a679d4d6 Mon Sep 17 00:00:00 2001 Subject: [PATCH 9/9] New SDL patch to support blinking the display
This uses low-level EPD control to wait for display update to complete,
this is required as otherwise drawing black then white produces tearing.
It's a bit ugly really with use of direct call from the generic driver to the fb driver.


The patch also ensures the update scheme is set to UPDATE_SCHEME_QUEUE
and forces monochrome mode.